test/LayoutTests/Panel/lastChangeKO.html

<html>

<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/ChannelPlate.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/RemoteMethodCall.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/AsyncMachine.js"></script>
<script src="../../../DebugLogger.js"></script>
<script src="../PatientSelector.js"></script>
<script src="../panel-test.js"></script>

<script>

window.addEventListener('load', function() {
  testScript();
});

function test_lastChange() {
  console.log("------------test begins------------");
  AsyncMachine.pushOp(evaluateInPage, 'window.location="http://localhost:8686/test/kodemo/qpd.html"');
  AsyncMachine.pushOp(openQuerypointPanel);
  AsyncMachine.pushOp(openSourceFile, 'demo.js');
  
  var objPropSelector = [
      {type: 'span.cm-variable', text:'bar'},  // call in onClick
      {type: 'span.cm-variable', text:'bar'},  // func decl
      {type: 'span.cm-property', text: 'value'},  // (!condition.value)
      {type: 'span.cm-property', text: 'value'}   // value  in condition.value
    ];
  AsyncMachine.pushOp(selectTokenInSource, objPropSelector);
  AsyncMachine.pushOp(verifyTokenView, 'condition.value');
  AsyncMachine.pushOp(clickQPOperation, 'lastChange');
  AsyncMachine.pushOp(whenSelectorAll, '.currentLoadNumber', '1');
  AsyncMachine.pushOp(evaluateInPage, 'document.querySelector("#myButton").dispatchEvent(new MouseEvent("click"))');
  AsyncMachine.pushOp(whenSelectorAll, 'td.activation', '1:7:0');
  AsyncMachine.pushOp(extractText, 'td.traceValue');
  AsyncMachine.pushOp(extractAttr, 'span.command', 'data-url');

  AsyncMachine.runOp();
}

var testScript = test_lastChange;
</script>
<body>
<h1>Feature Test lastChange</h1>
</body>
</html>
