test/LayoutTests/Panel/lastChangeKO-expected.txt

http://localhost:8686/test/kodemo/qpd.html
value in   if (!condition.value) {
true
undefined
http://localhost:8686/test/kodemo/demo.js?start=572&end=580&
